Update README.md
An expensive restaurant in the center of Antwerpen with savory meals. If you are a chicken lover, must go and taste the menu of Hector Chicken where a variety of roosters with delicious sauces and flavorful desserts are reachable for everyone. But remember it is a little costly cafeteria. Although, this is the best cafe for chicken lovers to dine in the savor.
Location 
The Centrum area is the most appropriate areas for several small restaurants with dissimilar refreshments. Hector Chicken is also located in the center area near the train station. A lot of civilians, visitors, and especially chicken lovers go there, and devour the meals of this cafeteria.



Web https://artisteology.com/blogs/chicken-lovers-blogger/ 
Blgr https://artisteology.blogspot.com/2024/08/chicken-lovers.html 
Whtspp https://whatsapp.com/channel/0029VafYJegHVvTWh5QaKD2z 
Thread  https://www.threads.net/@dailyroutineineurope 

#blog #blogger #blogging #artisteology #art #artist #design #antwerpen #belgium #europe #europetravel #quora #writter #contentwriter #contentcreator